requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-docx
PyMuPDF
llama-cpp-python
pydantic
python-multipart
